+++ src/main/java/edu/uga/miage/m1/polygons/gui/shapes/SimpleShape.java
@@ -18,7 +18,7 @@ public interface SimpleShape
      **/
     void draw(Graphics2D g2);
     
-    float getX();
+    int getX();
     
-    float getY();
+    int getY();
 }
